Removed logic in main to a seperate class
